% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{add_text_variables}
\alias{add_text_variables}
\title{Add readable variable names to SCB dataframe}
\usage{
add_text_variables(dfli, variable_list)
}
\arguments{
\item{dfli}{List of dataframes from SCB}

\item{variable_list}{SCB metadata list}
}
\description{
Add readable variable names to SCB dataframe
}
\keyword{internal}
