% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_if_partition_needed}
\alias{check_if_partition_needed}
\title{Check if SCB query should be partitioned}
\usage{
check_if_partition_needed(
  query,
  variable_list,
  treshold = 1e+05,
  out_queries = list()
)
}
\arguments{
\item{query}{Query list}

\item{variable_list}{SCB variable list}

\item{treshold}{Threshold for max combinations}

\item{out_queries}{Recursive accumulator}
}
\description{
Check if SCB query should be partitioned
}
\keyword{internal}
