% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{fetch_data_from_query}
\alias{fetch_data_from_query}
\title{Fetch data from SCB query}
\usage{
fetch_data_from_query(url, query)
}
\arguments{
\item{url}{API endpoint}

\item{query}{Query list}
}
\description{
Fetch data from SCB query
}
\keyword{internal}
