% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opposite_url.R
\name{get_api_url}
\alias{get_api_url}
\title{Helper function (not exported)}
\usage{
get_api_url(gui_url)
}
\arguments{
\item{gui_url}{A string value URL to a SCB table.}
}
\value{
The corresponding url to the API table
}
\description{
Helper function (not exported)
}
\keyword{internal}
