% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_full_query}
\alias{get_full_query}
\title{Create full SCB query}
\usage{
get_full_query(
  url,
  variable_list,
  print_query = FALSE,
  max_filter_values = 0,
  write_clipboard = TRUE
)
}
\arguments{
\item{url}{API endpoint}

\item{variable_list}{List of variables from SCB API}

\item{print_query}{Logical, whether to print the query}

\item{max_filter_values}{Integer, maximum number of values to filter}

\item{write_clipboard}{Logical, write query to clipboard}
}
\description{
Create full SCB query
}
\keyword{internal}
