% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opposite_url.R
\name{get_gui_url}
\alias{get_gui_url}
\title{Helper function (not exported)}
\usage{
get_gui_url(api_url)
}
\arguments{
\item{api_url}{A string value URL to a SCB table.}
}
\value{
The corresponding url to the GUI table
}
\description{
Helper function (not exported)
}
\keyword{internal}
