% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opposite_url.R
\name{get_opposite_url}
\alias{get_opposite_url}
\title{Send the URL to a SCB table.}
\usage{
get_opposite_url(url, write_clipboard = TRUE)
}
\arguments{
\item{url}{A string value URL-adress to a SCB table.}
}
\value{
A string value URL to the corresponding GUI or API table.
}
\description{
Get the URL to the corresponding GUI or API table.
}
