% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query.R
\name{get_query}
\alias{get_query}
\title{Download data from SCB API}
\usage{
get_query(url, max_filter_values = 0)
}
\arguments{
\item{url}{SCB API or GUI URL}

\item{max_filter_values}{Limit number of values per filter}
}
\value{
A tibble with SCB data
}
\description{
Download data from SCB API
}
