% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scb_data.R
\name{get_scb_data}
\alias{get_scb_data}
\title{Download data from SCB API}
\usage{
get_scb_data(
  url,
  query_li = list(),
  print_query = FALSE,
  max_filter_values = 0,
  write_clipboard = TRUE
)
}
\arguments{
\item{url}{SCB API or GUI URL}

\item{query_li}{Optional query list}

\item{print_query}{Print and copy R code to clipboard}

\item{max_filter_values}{Limit number of values per filter}

\item{write_clipboard}{Copy output to clipboard}
}
\value{
A tibble with SCB data
}
\description{
Download data from SCB API
}
