% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_variable_list}
\alias{get_variable_list}
\title{Get variable list from SCB}
\usage{
get_variable_list(url)
}
\arguments{
\item{url}{URL to SCB API}
}
\description{
Get variable list from SCB
}
\keyword{internal}
