% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{manage_function_list_query}
\alias{manage_function_list_query}
\title{Helper functions (not exported)}
\usage{
manage_function_list_query(list_query)
}
\arguments{
\item{list_query}{A list containing query elements.}
}
\description{
Helper functions (not exported)
}
\keyword{internal}
