% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opposite_url.R
\name{process_url}
\alias{process_url}
\title{Helper function (not exported)}
\usage{
process_url(single_url)
}
\arguments{
\item{single_url}{A string value URL to a SCB table.}
}
\description{
Helper function (not exported)
}
\keyword{internal}
